Predator_Skeleton/scraper_orchestrator/requirements.txt
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.4.26
charset-normalizer==3.4.2
chromedriver==2.24.1
h11==0.16.0
idna==3.10
kafka-python==2.0.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
schedule==1.2.1
selenium==4.21.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
wsproto==1.2.0
